Update - Problem with cnls fit, compare the gui and the test code

+++ src/GUI/Utils/eis_plots.py
@@ -1,6 +1,7 @@
 import os
 import numpy as np
 import dearpygui.dearpygui as dpg
+import src.GUI.Utils as gui_utils
 
 def update_single_plots(config):
     """
@@ -16,7 +17,7 @@ def update_single_plots(config):
     for idx, data_category in enumerate(["KK_data", "truncated", "LCcorrect", "smooth",  "extrapolation"]):
         dpg.delete_item(f"tab_eis_{data_category}_plot_single")
         with dpg.tab(label=data_category, tag=f"tab_eis_{data_category}_plot_single", parent="tab_bar_eis_plot_single"):
-            if config.display_file != [] and config.display_file is not None and os.path.splitext(config.display_file)[0] in config.store.keys() and config.store[os.path.splitext(config.display_file)[0]]['EIS'].KK_data['f'] is not None:
+            if config.display_file != [] and config.display_file is not None and os.path.splitext(config.display_file)[0] in config.store.keys() and 'EIS' in config.store[os.path.splitext(config.display_file)[0]] and config.store[os.path.splitext(config.display_file)[0]]['EIS'].KK_data['f'] is not None:
                 # Clear existing plots
                 dpg.delete_item(f"tab_eis_{data_category}_data_plot_single_KK")
                 dpg.delete_item(f"tab_eis_{data_category}_data_plot_single_Z_Phase")
@@ -128,7 +129,7 @@ def update_all_plots(config):
     for idx, data_category in enumerate(["KK_data", "truncated", "LCcorrect", "smooth",  "extrapolation"]):
         dpg.delete_item(f"tab_eis_{data_category}_plot_all")
         with dpg.tab(label=data_category, tag=f"tab_eis_{data_category}_plot_all", parent="tab_bar_eis_plot_all"):
-            if config.display_file != [] and config.display_file is not None and os.path.splitext(config.display_file)[0] in config.store.keys() and config.store[os.path.splitext(config.display_file)[0]]['EIS'].KK_data['f'] is not None:
+            if config.display_file != [] and config.display_file is not None and os.path.splitext(config.display_file)[0] in config.store.keys() and 'EIS' in config.store[os.path.splitext(config.display_file)[0]] and config.store[os.path.splitext(config.display_file)[0]]['EIS'].KK_data['f'] is not None:
                 # Clear existing table rows
                 dpg.delete_item(f"tab_eis_{data_category}_data_plot_all_KK")
                 dpg.delete_item(f"tab_eis_{data_category}_data_plot_all_Z_Phase")
@@ -166,12 +167,12 @@ def update_all_plots(config):
                         file_name_no_ext = os.path.splitext(file_name)[0]
                         data = config.store[file_name_no_ext]['EIS']
                         # KK results
-                        dpg.add_scatter_series(data['KK_data']['f'], data['KK_data']['delta_Re_kk'], parent=y_axis_KK, label=f"Re-{file_name_no_ext}")
-                        dpg.add_scatter_series(data['KK_data']['f'], data['KK_data']['delta_Im_kk'], parent=y_axis_KK, label=f"Im-{file_name_no_ext}")
+                        dpg.add_scatter_series(data['KK_data']['f'], data['KK_data']['delta_Re_kk'], parent=y_axis_KK, label=gui_utils.small_functions.string_abbreviation(f"Re-{file_name_no_ext}", 13, 12))
+                        dpg.add_scatter_series(data['KK_data']['f'], data['KK_data']['delta_Im_kk'], parent=y_axis_KK, label=gui_utils.small_functions.string_abbreviation(f"Im-{file_name_no_ext}", 13, 12))
 
                         # Impedance module and phase
-                        dpg.add_scatter_series(data['raw']['f'], abs(data['raw']['Z']), parent=y_axis1_Z_Phase, label=f"Z-{file_name_no_ext}")
-                        dpg.add_scatter_series(data['raw']['f'], -np.degrees(np.angle(data['raw']['Z'])), parent=y_axis2_Z_Phase, label=f"Phase-{file_name_no_ext}")
+                        dpg.add_scatter_series(data['raw']['f'], abs(data['raw']['Z']), parent=y_axis1_Z_Phase, label=gui_utils.small_functions.string_abbreviation(f"Z-{file_name_no_ext}", 12, 12))
+                        dpg.add_scatter_series(data['raw']['f'], -np.degrees(np.angle(data['raw']['Z'])), parent=y_axis2_Z_Phase, label=gui_utils.small_functions.string_abbreviation(f"Phase-{file_name_no_ext}", 16, 12))
 
                     dpg.add_plot_legend(parent=f"tab_eis_{data_category}_data_plot_all_KK")
                     dpg.add_plot_legend(parent=f"tab_eis_{data_category}_data_plot_all_Z_Phase")
@@ -211,11 +212,11 @@ def update_all_plots(config):
                     for file_name in config.selected_files:
                         file_name_no_ext = os.path.splitext(file_name)[0]
                         data = config.store[file_name_no_ext]['EIS']
-                        dpg.add_line_series(data[data_category]['f'], data[data_category]['Re'], parent=y_axis_Re, label=f"{data_category[0].upper() + data_category[1:]}-{file_name_no_ext}")
+                        dpg.add_line_series(data[data_category]['f'], data[data_category]['Re'], parent=y_axis_Re, label=gui_utils.small_functions.string_abbreviation(f"{data_category[0].upper() + data_category[1:]}-{file_name_no_ext}", 12, 12))
 
-                        dpg.add_line_series(data[data_category]['f'], -data[data_category]['Im'], parent=y_axis_Im, label=f"{data_category[0].upper() + data_category[1:]}-{file_name_no_ext}")
+                        dpg.add_line_series(data[data_category]['f'], -data[data_category]['Im'], parent=y_axis_Im, label=gui_utils.small_functions.string_abbreviation(f"{data_category[0].upper() + data_category[1:]}-{file_name_no_ext}", 12, 12))
 
-                        dpg.add_line_series(data[data_category]['Re'], -data[data_category]['Im'], parent=y_axis_Nyquist, label=f"{data_category[0].upper() + data_category[1:]}-{file_name_no_ext}")
+                        dpg.add_line_series(data[data_category]['Re'], -data[data_category]['Im'], parent=y_axis_Nyquist, label=gui_utils.small_functions.string_abbreviation(f"{data_category[0].upper() + data_category[1:]}-{file_name_no_ext}", 12, 12))
 
                     dpg.add_plot_legend(parent=f"tab_eis_{data_category}_data_plot_all_Re")
                     dpg.add_plot_legend(parent=f"tab_eis_{data_category}_data_plot_all_Im")
